스프링핵심원리-기본편/core/src/test/java/hello/core/dicount/RateDiscountPolicyTest.java
package hello.core.dicount;

import hello.core.member.Grade;
import hello.core.member.Member;
import org.junit.jupiter.api.DisplayName;
import org.junit.jupiter.api.Test;

import static org.assertj.core.api.Assertions.assertThat; //이 방식으로써야함!!!!
class RateDiscountPolicyTest { //정률할인정책이 실제로 10%할인 되는지 검증하기

    RateDiscountPolicy discountPolicy = new RateDiscountPolicy();

    @Test
    @DisplayName("VIP는 10% 할인이 적용되어야 한다") //테스트 좌측 창에 표기될 테스트 이름
    void vip_o() {
        //given
        Member member = new Member(1L, "memberVIP", Grade.VIP);
        //when
        int discount = discountPolicy.discount(member, 10000);
        //then
        assertThat(discount).isEqualTo(1000);
    }

    @Test
    @DisplayName("VIP가 아니면 할인이 적용되지 않아야한다.") //테스트 좌측 창에 표기될 테스트 이름
    void vip_x()
    {
        //given
        Member member = new Member(1L, "memberBasic", Grade.BASIC);
        //when
        int discount = discountPolicy.discount(member, 10000);
        //then
        assertThat(discount).isEqualTo(0);
    }

}